shiba_file_name function test
