+++ test/shiba_test.py
@@ -1,30 +1,11 @@
 
-from app.shiba_app import get_advice
-import random
-
-# TODO: test the code
-
-# def test_format_usd():
-#     #result = format_usd(1999.99)
-#     assert format_usd(1999.99) == "$1,999.99"
-#
-# mock_products_filepath = os.path.join(os.path.dirname(__file__), "mock_data", "mock_products.csv")
-# mock_products_df = read_csv(mock_products_filepath)
-# mock_products = mock_products_df.to_dict("records")
-#
-# def test_find_product():
-#     #with valid product id, returns the product info
-#     valid_result = find_product("8", mock_products)
-#     assert valid_result =={
-#         'aisle':'Aisle C',
-#         'department':'snacks',
-#         'id': 8,
-#         'name': 'Product 8',
-#         'price': 10.0
-#     }
-#     #with invalid product id, returns None:
-#     invalid_result = find_product("999999", mock_products)
-#     assert invalid_result == None
+from app.shiba import shiba_file_name
 
 
-def test_get_advice():
+advicestr = "avocado"
+shibalink = "https://cdn.shibe.online/shibes/c64a387d249909f280ebf3b2cb13b352c1539882.jpg"
+
+
+def test_shiba_file_name():
+    assert shiba_file_name("random",shibalink,advicestr) == "random_c64a387d249909f280ebf3b2cb13b352c1539882.jpg"
+    assert shiba_file_name("ask",shibalink,advicestr) == "avocado_c64a387d249909f280ebf3b2cb13b352c1539882.jpg"
